+++ lib/src/model/field.dart
@@ -17,24 +17,17 @@ class Field {
 
   String? comment;
 
-  static const Set<String> innerDartTypes = {
-    "int",
-    "bool",
-    "double",
-    "String",
-    "Map",
-    "List"
-  };
+  static const Set<String> innerDartTypes = {"int", "bool", "double", "String", "Map", "List"};
 
   Field(
       {required this.name,
       required this.type,
       required this.isOptional,
-      required this.isFinal})
+      required this.isFinal,
+      this.expectEntities = const []})
       : valueType = null,
         keyType = null,
-        isPlain = true,
-        expectEntities = [];
+        isPlain = true;
 
   Field.map(
       {required this.name,
@@ -61,8 +54,7 @@ class Field {
 
   bool get isCustomType => !innerDartTypes.contains(type);
 
-  bool get isValueCustomType =>
-      (!isValueTypeDynamic) && (!innerDartTypes.contains(valueType!));
+  bool get isValueCustomType => (!isValueTypeDynamic) && (!innerDartTypes.contains(valueType!));
 
   bool get isValueTypeDynamic => valueType == 'dynamic';